test: LottoResultTest 작성
